Bug 1568 - AccuApproval - When no default statuses are selected in admin, and create new app it still brings in a default analysis status
Added a condition to check if there is a value for the "Analysis Status" then it will display the value.

This took a while because the IsNull was not returning a correct value 100% of the time.  Now the results are 100% accurate.

I added code to program "incluce/loanmaintappadmin.asp" to determine if the value is NULL or a key.  [Line 151 - 155]

James