Bug 1539 - Editors Able to See Employee File Documents
I modifed the processing so that if the customer is an "Employee" and the user does not have "Access to Employee Files" and they press the "SCAN" it will not allow them to scan to the same folder were documents are. I will how ever do a check to see if there are documents in the folder and if there are none it will scan the documents to that folder of not it will create a new folder as "Scan into New Document".

Also, the following buttons will not be displayed: MOVE, COPY, DELETE, REPLACE, MERGE and REPLACE if the above condition is meet.

I modified the following programs:

a) include/cust_creditdocbody.inc - I modifed the lines of code to pass more data to the documentmaint.asp page from the URL and not read the database again. [lines 346 - 350] [lines13 -14] [line 332 & 359]

b) include/display_customer_across.inc - I added code to set a session variable if the customer is an employee "Session("IsEmployee") [line 33]

c) documentmain.asp - This program does the processing for the buttons that are to be hidden and the "SCAN" button. [lines 468 - 480] [lines 605, 618, 632, 675 - 685, 696 -719, 764, 788]
James Muschette8/27/2014 10:17 AM
d) include/cust_creditdocbody.inc - to be able to pass data in the URL for from the program "documentaction.asp" [Lines 259 - 268, 275, 281]

e) documentaction.asp - the code here is to not allow the display of the option "Merge Files" if the condition is valid as stated above.

James