Bug Fix for physical upload paths stored in cookies.
Changes for the upload navigation saves the user's current path as a cookie. The issue is that previous bugs left the path as a physical path instead of a logical/relative path. The code adds an error check so that program can continue properly as later bullet proofing will reset the path to a relative one.