Bug #1582 - DocumentScheduler never finishes execution.
This was discovered at the State Bank of Fairbault. After troubleshooting I discovered that the user defined unit to increment the date by allows 0 or negative numbers. This should not be allowed because it allows the program to get stuck in an infinite loop. The DocumentScheduler has been modified to add checks for non-positive unit increments and log them to the output file.