Bug #1579 - Factoring cross collateral issue
The issue was due to the scanned_image path property not being valid. I've added an extra check on the path before refactoring starts to log a message indicating the path is bad. Once logged, exception processing will continue skipping refactoring since the path is bad.

Also added was extra logging and small change to exception processing to handle exceptions that are set or pending but the date is NULL. It will update the time stamp as best as it can with the information provided, otherwise it will use the processing date.