Refactor the AccuAccount.Web.Exception and AccuAccount.Web.ExceptionService projects for consistency.
Also during the refactor, the loading of the db.config file has been moved to appropriate applications (e.g. Web for the Service and ConsoleApp for the nightly process). The path to the config file is determined differently when Web vs. ConsolApp. So the ExceptonService is now responsible for getting the config path and passing to ExceptionValidator. The ConsoleApp the same.