1596 | Merging causes tmp files Version 13
A new C# class was created entitled "BinaryFileCheck" containing 2 methods (IsDocumentTruePDF and IsDocumentTrueTIFF) that read the binary headers of the files to identify if the file actually is what is indicated in the file extension. These were added to the "AccuAccount.Web" Namespace and are accessed as such. Now, before the merge is conducted, these methods validate all files that are involved in the merge and will return an error if a file is mis-labelled. This validation code was added to uploadmaint.aspx and documentmerge.aspx.