Bug #1167 - upload from co-borrower instance assigns documents and files to the co-borrower and not the primary borrower.
The fix was to modify the documentaction.asp page to check for primaryCustomerId based on the loan table instead of the passed value from the customer page. I put the fix here as it was easier to check at this point then change the potential links in different pages. If the loanId passed in is not blank then a check for the primaryCustomerId will be done and if it is different then the customerId passed in it will use the primaryBorrowerId instead. This also fixes a past issue with DocumentTabs (documentTab table) being created where the (customerId,loanId) value pair do not match the loan table pair.