Bug #1677 - Audit displays documents that it shouldn't
This issue had to do with differences between the SQL of the auditbuild.asp and the customer.asp page. I've modified the SQL in auditbuild.asp to be the same as in the customer.asp page. auditbuild.asp factored in the documentTab.docTabDocumentStatusType in it's calculation but customer.asp doesn't so I removed it as it was inconsistent data.