Bug #1649 - Approval frame not showing list of pending approvals
When approvals were done an assumption was made that the first  or primary field of each record was a Text field. In a particular case a Module was setup to have a choiceField as the primary field of a record and nothing was displayed as it wasn't a text field. The fix was to add the extra logic necessary to build the SQL based on the fieldType (text, choice, date, currency or memo) of the first field of the record. Once the SQL was changed to select on the correct flex field table the data appeared.