Bug #1675 - Updating custom exception to "sync" the assigned user and default assigned user.
Not strictly a bug but this change makes maintenance of exceptions easier. Normally the assignedUser and the defaultAssignedUser can be different. The maintenance pages have been modified to keep the assignedUser of the Exception the same as the defaultAssignedUser of the ExceptionDefinition. This is only done for Custom Exceptions at the Customer/Account/Collateral as their is a one-to-one relationship between the Exception and its Definition.