Bug #1694 - Deleting users generates an error
The error was due to a problem with writing to the activity log. The solution though for this ticket is that users shouldn't be deleted as there may be database references associated with them, particularly approval history and admin maintenance summary. The user maintenance pages have been modified to display active vs. inactive users like AccuAccount. Inactive users will be prevented from logging in but their profile information will remain so there is no broken data references in the database.