Merged revision(s) 5374, 5377, 5379, 5387 from AccuSystems/Products/Development/AccuLoan/Apps/Web/branches/v13.0.36: # 1032: delete hardcorded "Primary Borrower" from SQL query ........ 1539 | Editors Able to See Employee File Documents Version 13
Made a modification to the document permissions to tighten the security when accessing an uploaded document that is an employee file.
........
1539 | Editors Able to See Employee File Documents
Version 13

Corrected typo "docHidEmployeefileYN" to "docHideEmployeefileYN" in document permissions.
........
1539 | Editors Able to See Employee File Documents
Version 13

The variable 'employeeViewer' needs to be reset to 'False' as the 'True' was being carried over from the credit side affecting the account document permissions.
........