Bug #1696 - FNB Sioux Falls - Document Images/Titles Not Correct After Cross Collateralizing
The issue had occurred due to how the copy of the files occurred. It relied on the source and target document record information to copy the files over. The problem was due to the SQL not constrained enough as it was written when the rule was one Document to one DocumentTab. When we expanded DocumentTabs to allow multiple documents the SQL code was missed. The effect has been that only one document from each tab is copied over to all the document records in the tab. Adding the extra constraint clears up the problem and all files are copied to their respective target document records.

The current constraint was just on documentType and documentSubType. This allowed all combinations of source documents to target documents. By adding the filename, which is unique for each folder, this prevents the combination data causing one file to keep overwriting files in the target documentTab folders.