Bug #1692 - BoBV document search filtering issues
refactored the document search SQL generation for clarity. Small changes to the code over time have resulted in inconsistent SQL that create bad combinations of joins on Flex Fields that it shouldn't have. For example a document type that had a Text based flex field was joined on the Choice based flex field table which would cause no results to be returned.