Merged revision(s) 5563-5566 from AccuSystems/Products/Development/AccuLoan/Apps/Web/branches/v13.0.37: 1669 | Orphaned Documents are Reporting Documents that are Not Truly Orphaned Version 13
Made changes to the orphaned document so that if a document definition is not found because a change of type, causing a renaming of the file directory, the software will use the defaultName only to find the document definition and then its extracts the correct document definition using the documentSubTypeId, documentTypeId, and loanTypeId of the wrong document Definition to get us the correct document definition.
........
1669 | Orphaned Documents are Reporting Documents that are Not Truly Orphaned
Version 13

Changed the methods "GetDocumentFromAccountNumberAndLoanFileAndCustomerNumberAndFileName" and "GetDocumentDefinitionByDefaultName" to use FirstOrDefault to prevent an exception when there are multiple records.
........
Add RemoveTrailingSlash function to URL in order to make it work in any server no matter what address is store as acculoan.serverUrl property

........

Also rebuilt the project.. PaulJ