Bug 1648 - Converting cross collateral points to itself Merged from v13 development branch
Due to a syntax error (the name of the source collateral loan id variable) was assumed to always be blank and therefore caused the collateral being converted into a cross collateral to show on the potential list of collaterals. The code is fixed to ensure the collateral being converted does not show up now.