Merged revision(s) 5647-5648 from AccuSystems/Products/Development/AccuLoan/Apps/Web/trunk: Merging in Merged revision(s) 5630-5631 from AccuSystems/Products/Development/AccuLoan/Apps/Web/branches/v13.0.37: Bug #1739 - Ambiguous column name in Audit search
The SQL statement to build the search included multiple copies of the collateralSequence field when I updated the SQL to use the new viewAccountsAndCollaterals. I've removed the extra copies of the field so the ambiguous field error does not occur 
........
Bug #1739 - Ambiguous column name in Audit search

The SQL statement to build the search included multiple copies of the collateralSequence field when I updated the SQL to use the new viewAccountsAndCollaterals. I've removed the extra copies of the field so the ambiguous field error does not occur.

Made sure the credit collateral sequence is the same as the account collateral sequence of -1.
........


........