Merged revision(s) 5719-5720 from AccuSystems/Products/Development/AccuLoan/Apps/Web/branches/v14.0.25: 1767 | Loading Orphaned Documents page produces error Version 15
*  Added an additional "if else" statement when the documentSubTypeId is null, so that it marks the document as unmapped if the documentSubTypeId is missing. This will prevent the error reported in the ticket.

*  Changed the version of the target framework of the website project to .NET Framework 4.5.1
........
Bug #1761 - Converting to cross collateral causes error

The issue is due to the order deletes occurring for the collateral being converted. The FK constraint on the exceptedDocument table is enabled in the staging database so I've reordered the deletes to cascade in the order necessary not to trigger FK constraints.
........