1741 | Issue when in Existing Document Search when Searching by Customer
The search will always fire the sub routine called "ProcessQuickSearch" which looks for 4 distinct fields "searchloan", "searchName", "searchCustomerNumber", and "searchTaxId". If one of these is found, in this case "searchName" the subroutine fires and finds the named customer and goes directly to the customer page without considering any other of the search parameters. I added in an "If" statement to check for other search parameters.  If it finds one, it will exit the sub routine and do a full search using all the available parameters.