Merged revision(s) 5598, 5768, 5772, 5784, 5788 from AccuSystems/Products/Development/AccuLoan/Apps/Web/branches/v13.0.37: Bug 1648 - Converting cross collateral points to itself
Due to a syntax error (the name of the source collateral loan id variable) was assumed to always be blank and therefore caused the collateral being converted into a cross collateral to show on the potential list of collaterals. The code is fixed to ensure the collateral being converted does not show up now.
........
Bug #1762 - CBC Issue Moving documents

The ASPX upload maintenance page has been modified to better handle issues with intermittent file locking by gracefully exiting the process and warning the user.
........
Bug #1768 - Document Cover Sheets display wrong barcode

The Credit and Loan sections did not use the appropriate document barcode string and has been fixed. The collateral was correct and so no change was necessary.
........
Bug #1783 - Unable to Delete Account or Collateral Records

Added necessary deletes to cascade in the correct order to ensure records are removed that do not violate the FK constraints of the database.
........
Updating the version numbering and the DBUpdate.cmd for the new 13.0.36.2 hotfix packaging.
........