Merged revision(s) 5801 from AccuSystems/Products/Development/AccuLoan/Apps/Web/branches/v13.0.37: 1774 | Readers are not able to email documents from AccuAccount/domain permissions are the issue Version 13 and 15
Created a new function to normalize the path slashes depending on the the type of path being passed.  This fixes the issue in the ticket as some clients had their "fullpath" wrongly configured with forward slashes causing the folder to not be correctly identified causing the new directory not to be created.
........